Extract BookingEntity-to-ViewState mapping into WooPosBookingViewStateMapper
