+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/bookings/WooPosBookingsViewModel.kt
@@ -18,25 +18,17 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.bookings.BookingCustomerInfo
-import org.wordpress.android.fluxc.persistence.entity.BookingEntity
-import org.wordpress.android.fluxc.persistence.entity.isAttendanceStatusEditable
-import java.math.BigDecimal
-import java.time.Duration
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
-import java.time.format.FormatStyle
 import javax.inject.Inject
 
 @HiltViewModel
 class WooPosBookingsViewModel @Inject constructor(
     private val bookingListHandler: BookingListHandler,
     private val dateTimeProvider: DateTimeProvider,
+    private val mapper: WooPosBookingViewStateMapper,
 ) : ViewModel() {
 
     companion object {
         private const val MIN_LOADING_DURATION_MS = 300L
-        private const val MINUTES_PER_HOUR = 60
     }
 
     private val _state = MutableStateFlow<WooPosBookingsState>(WooPosBookingsState.Loading)
@@ -92,7 +84,8 @@ class WooPosBookingsViewModel @Inject constructor(
                 }
 
                 val items = bookings.associate { booking ->
-                    mapToItemViewState(booking) to mapToDetailsViewState(booking)
+                    mapper.mapToItemViewState(booking, selectedBookingId) to
+                        mapper.mapToDetailsViewState(booking)
                 }
 
                 val selectedDetails = selectedBookingId?.let { id ->
@@ -237,171 +230,5 @@ class WooPosBookingsViewModel @Inject constructor(
                 // TBD: handle email receipt
             }
         }
-    }
-
-    private fun mapToItemViewState(booking: BookingEntity): WooPosBookingsState.BookingItemViewState {
-        return WooPosBookingsState.BookingItemViewState(
-            id = booking.id.value,
-            title = booking.order.productInfo?.name ?: "#${booking.id.value}",
-            date = booking.start.toString(),
-            total = booking.order.paymentInfo?.total?.toPlainString() ?: "",
-            customerEmail = booking.order.customerInfo?.billingEmail?.ifBlank { null },
-            isSelected = booking.id.value == selectedBookingId,
-            status = mapBookingStatus(booking.status),
-            statusSlug = booking.status.key,
-            createdAtMillis = booking.start.toEpochMilli()
-        )
-    }
-
-    private fun mapToDetailsViewState(
-        booking: BookingEntity
-    ): WooPosBookingsState.BookingDetailsViewState {
-        val zone = ZoneId.systemDefault()
-        val dateFormatter = DateTimeFormatter.ofLocalizedDate(FormatStyle.FULL).withZone(zone)
-        val timeFormatter = DateTimeFormatter.ofLocalizedTime(FormatStyle.SHORT).withZone(zone)
-
-        val bookingName = booking.order.productInfo?.name ?: "#${booking.id.value}"
-        val appointmentTime = "${timeFormatter.format(booking.start)} - ${timeFormatter.format(booking.end)}"
-
-        val customerInfo = booking.order.customerInfo
-        val customerName = listOfNotNull(
-            customerInfo?.billingFirstName,
-            customerInfo?.billingLastName
-        ).joinToString(" ").ifBlank { null }
-
-        val headerSubtitle = buildString {
-            append(bookingName)
-            customerName?.let { append(" \u00B7 $it") }
-        }
-
-        return WooPosBookingsState.BookingDetailsViewState(
-            id = booking.id.value,
-            number = "#${booking.id.value}",
-            status = mapBookingStatus(booking.status),
-            actionsState = WooPosBookingsState.BookingActionsState.Loaded(
-                listOf(WooPosBookingsState.BookingAction.EmailReceipt(booking.orderId))
-            ),
-            headerTitle = appointmentTime,
-            headerSubtitle = headerSubtitle,
-            attendanceBadge = mapAttendanceBadge(booking.attendanceStatus),
-            bookingName = bookingName,
-            appointmentDate = dateFormatter.format(booking.start),
-            appointmentTime = appointmentTime,
-            duration = formatDuration(Duration.between(booking.start, booking.end)),
-            teamMember = null,
-            location = null,
-            customerSection = buildCustomerSection(customerInfo, customerName, booking.customerNote),
-            attendanceSection = buildAttendanceSection(booking),
-            paymentSection = buildPaymentSection(booking),
-            bookingNote = booking.note.ifBlank { null },
-        )
-    }
-
-    private fun buildAttendanceSection(
-        booking: BookingEntity
-    ): WooPosBookingsState.AttendanceSection? {
-        if (!booking.isAttendanceStatusEditable) return null
-        val selection = when (booking.attendanceStatus) {
-            BookingEntity.AttendanceStatus.CheckedIn -> WooPosBookingsState.AttendanceState.ATTENDED
-            BookingEntity.AttendanceStatus.NoShow -> WooPosBookingsState.AttendanceState.UNATTENDED
-            else -> null
-        }
-        return WooPosBookingsState.AttendanceSection(selection = selection)
-    }
-
-    private fun buildPaymentSection(
-        booking: BookingEntity
-    ): WooPosBookingsState.PaymentSection {
-        val paymentInfo = booking.order.paymentInfo
-        val discount = paymentInfo?.let { it.total - it.subtotal } ?: BigDecimal.ZERO
-        val paymentTotal = paymentInfo?.let { it.total + it.totalTax } ?: BigDecimal.ZERO
-        val isPaid = booking.status == BookingEntity.Status.Paid ||
-            booking.status == BookingEntity.Status.Complete
-
-        return WooPosBookingsState.PaymentSection(
-            serviceAmount = paymentInfo?.subtotal?.toPlainString() ?: "-",
-            taxAmount = paymentInfo?.totalTax?.toPlainString() ?: "-",
-            discountAmount = if (discount.compareTo(BigDecimal.ZERO) != 0) {
-                "-${discount.abs().toPlainString()}"
-            } else {
-                "-"
-            },
-            totalAmount = paymentTotal.toPlainString(),
-            paidWithLabel = if (isPaid) paymentInfo?.paymentMethodTitle else null,
-            showPayButtons = !isPaid,
-        )
-    }
-
-    private fun buildCustomerSection(
-        customerInfo: BookingCustomerInfo?,
-        customerName: String?,
-        customerNote: String?,
-    ): WooPosBookingsState.CustomerSection? {
-        val email = customerInfo?.billingEmail?.ifBlank { null }
-        val phone = customerInfo?.billingPhone?.ifBlank { null }
-        val billingAddress = buildBillingAddress(customerInfo)
-        val note = customerNote?.ifBlank { null }
-
-        val hasContent = customerName != null || email != null || phone != null ||
-            billingAddress != null || note != null
-        if (!hasContent) return null
-
-        return WooPosBookingsState.CustomerSection(
-            name = customerName,
-            email = email,
-            phone = phone,
-            billingAddress = billingAddress,
-            note = note,
-        )
-    }
-
-    private fun formatDuration(duration: Duration): String {
-        val hours = duration.toHours()
-        val minutes = duration.toMinutes() % MINUTES_PER_HOUR
-        return when {
-            hours > 0 && minutes > 0 -> "$hours hr $minutes min"
-            hours > 0 -> "$hours hr"
-            else -> "$minutes min"
-        }
-    }
-
-    private fun buildBillingAddress(customerInfo: BookingCustomerInfo?): String? {
-        if (customerInfo == null) return null
-        val parts = listOfNotNull(
-            customerInfo.billingAddress1,
-            customerInfo.billingAddress2,
-            customerInfo.billingCity,
-            customerInfo.billingState,
-            customerInfo.billingPostcode,
-            customerInfo.billingCountry,
-        ).filter { it.isNotBlank() }
-        return parts.joinToString(", ").ifBlank { null }
-    }
-
-    private fun mapAttendanceBadge(
-        status: BookingEntity.AttendanceStatus
-    ): WooPosBookingsState.AttendanceState? {
-        return when (status) {
-            BookingEntity.AttendanceStatus.CheckedIn -> WooPosBookingsState.AttendanceState.ATTENDED
-            BookingEntity.AttendanceStatus.NoShow -> WooPosBookingsState.AttendanceState.UNATTENDED
-            else -> null
-        }
-    }
-
-    private fun mapBookingStatus(status: BookingEntity.Status): WooPosBookingStatus {
-        val colorKey = when (status) {
-            BookingEntity.Status.Complete -> WooPosBookingStatusColorKey.COMPLETED
-            BookingEntity.Status.Paid -> WooPosBookingStatusColorKey.COMPLETED
-            BookingEntity.Status.Confirmed -> WooPosBookingStatusColorKey.PROCESSING
-            BookingEntity.Status.PendingConfirmation -> WooPosBookingStatusColorKey.ON_HOLD
-            BookingEntity.Status.Unpaid -> WooPosBookingStatusColorKey.ON_HOLD
-            BookingEntity.Status.Cancelled -> WooPosBookingStatusColorKey.FAILED
-            BookingEntity.Status.InCart -> WooPosBookingStatusColorKey.OTHER
-            is BookingEntity.Status.Unknown -> WooPosBookingStatusColorKey.OTHER
-        }
-        return WooPosBookingStatus(
-            text = status.key.replaceFirstChar { it.uppercase() }.replace("-", " "),
-            colorKey = colorKey
-        )
     }
 }
